refactor: Query mapping
