+++ grc20-core/src/mapping/entity/delete_one.rs
@@ -1,0 +1,55 @@
+use crate::{block::BlockMetadata, error::DatabaseError, indexer_ids, mapping::Query};
+
+pub struct DeleteOneQuery {
+    neo4j: neo4rs::Graph,
+    block: BlockMetadata,
+    id: String,
+    space_id: String,
+    space_version: String,
+}
+
+impl DeleteOneQuery {
+    pub fn new(
+        neo4j: &neo4rs::Graph,
+        block: &BlockMetadata,
+        id: String,
+        space_id: String,
+        space_version: String,
+    ) -> Self {
+        Self {
+            neo4j: neo4j.clone(),
+            block: block.clone(),
+            id,
+            space_id,
+            space_version,
+        }
+    }
+}
+
+impl Query<()> for DeleteOneQuery {
+    async fn send(self) -> Result<(), DatabaseError> {
+        const QUERY: &str = const_format::formatcp!(
+            r#"
+            MATCH (e:Entity {{id: $entity_id}}) -[r:ATTRIBUTE {{space_id: $space_id, max_version: null}}]-> (:Attribute)
+            SET r.max_version = $space_version
+            SET e += {{
+                `{UPDATED_AT}`: datetime($block_timestamp),
+                `{UPDATED_AT_BLOCK}`: $block_number
+            }}
+            "#,
+            UPDATED_AT = indexer_ids::UPDATED_AT_TIMESTAMP,
+            UPDATED_AT_BLOCK = indexer_ids::UPDATED_AT_BLOCK,
+        );
+
+        let query = neo4rs::query(QUERY)
+            .param("entity_id", self.id)
+            .param("space_id", self.space_id)
+            .param("space_version", self.space_version)
+            .param("block_timestamp", self.block.timestamp.to_rfc3339())
+            .param("block_number", self.block.block_number.to_string());
+
+        self.neo4j.run(query).await?;
+
+        Ok(())
+    }
+}
