+++ grc20-core/src/mapping/relation/delete_one.rs
@@ -1,0 +1,56 @@
+use crate::{block::BlockMetadata, error::DatabaseError, indexer_ids, mapping::Query};
+
+pub struct DeleteOneQuery {
+    neo4j: neo4rs::Graph,
+    block: BlockMetadata,
+    relation_id: String,
+    space_id: String,
+    space_version: String,
+}
+
+impl DeleteOneQuery {
+    pub(super) fn new(
+        neo4j: neo4rs::Graph,
+        block: BlockMetadata,
+        relation_id: String,
+        space_id: String,
+        space_version: String,
+    ) -> Self {
+        DeleteOneQuery {
+            neo4j,
+            block,
+            relation_id,
+            space_id,
+            space_version,
+        }
+    }
+}
+
+impl Query<()> for DeleteOneQuery {
+    async fn send(self) -> Result<(), DatabaseError> {
+        // TODO: Add relation entity deletion
+        const QUERY: &str = const_format::formatcp!(
+            r#"
+                MATCH () -[r:RELATION {{id: $relation_id}}]-> ()
+                WHERE r.space_id = $space_id
+                AND r.max_version IS NULL
+                SET r.max_version = $space_version
+                SET r += {{
+                    `{UPDATED_AT}`: datetime($block_timestamp),
+                    `{UPDATED_AT_BLOCK}`: $block_number
+                }}
+            "#,
+            UPDATED_AT = indexer_ids::UPDATED_AT_TIMESTAMP,
+            UPDATED_AT_BLOCK = indexer_ids::UPDATED_AT_BLOCK,
+        );
+
+        let query = neo4rs::query(QUERY)
+            .param("relation_id", self.relation_id)
+            .param("space_id", self.space_id)
+            .param("space_version", self.space_version)
+            .param("block_timestamp", self.block.timestamp.to_rfc3339())
+            .param("block_number", self.block.block_number.to_string());
+
+        Ok(self.neo4j.run(query).await?)
+    }
+}
