+++ grc20-sdk/examples/space_schema.rs
@@ -1,5 +1,8 @@
 use futures::{pin_mut, StreamExt, TryStreamExt};
-use grc20_core::{mapping::{entity_node::EntityNodeRef, query_utils::QueryStream}, neo4rs, system_ids};
+use grc20_core::{
+    mapping::{entity::EntityNodeRef, query_utils::QueryStream, RelationEdge},
+    neo4rs, system_ids,
+};
 use grc20_sdk::models::{property, space};
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 
@@ -45,7 +48,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                     type_.id
                 );
 
-                let properties = property::get_outbound_relations::<EntityNodeRef>(
+                let properties = property::get_outbound_relations::<RelationEdge<EntityNodeRef>>(
                     &neo4j,
                     system_ids::PROPERTIES,
                     type_.id,
@@ -75,7 +78,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                             .await?
                             .map(|triple| triple.value.value);
 
-                            let value_type = property::get_outbound_relations::<EntityNodeRef>(
+                            let value_type = property::get_outbound_relations::<RelationEdge<EntityNodeRef>>(
                                 &neo4j,
                                 system_ids::VALUE_TYPE_ATTRIBUTE,
                                 &property.to,
