+++ grc20-sdk/src/models/space/space_types_query.rs
@@ -1,13 +1,11 @@
 use futures::{Stream, TryStreamExt};
 
 use grc20_core::{
-    error::DatabaseError,
-    mapping::{
-        entity_node, prop_filter,
+    entity, error::DatabaseError, mapping::{
+        prop_filter,
         query_utils::{QueryStream, TypesFilter},
         EntityFilter, EntityNode, PropFilter, Query,
-    },
-    neo4rs, system_ids,
+    }, neo4rs, system_ids
 };
 
 use super::ParentSpacesQuery;
@@ -55,7 +53,7 @@ impl Query<Option<EntityNode>> for FindSpaceTypeQuery {
         }
 
         // Find all entities that have a TYPES relation to the Type entity
-        let mut results = entity_node::find_many(&self.neo4j)
+        let mut results = entity::find_many::<EntityNode>(&self.neo4j)
             .with_filter(
                 EntityFilter::default()
                     .id(prop_filter::value(self.id.clone()))
@@ -136,7 +134,7 @@ impl QueryStream<EntityNode> for FindSpaceTypesQuery {
         }
 
         // Find all entities that have a TYPES relation to the Type entity
-        let mut query = entity_node::find_many(&self.neo4j)
+        let mut query = entity::find_many::<EntityNode>(&self.neo4j)
             .with_filter(
                 EntityFilter::default()
                     .relations(TypesFilter::default().r#type(system_ids::SCHEMA_TYPE))
